Refactor json tests to align with new code
